fix: SessionsControllerをApi::V2名前空間に移動し、API環境でCookieを有効化
- SessionsControllerをapp/controllers/api/v2/に移動してApi::V2名前空間を追加
- ApplicationControllerにActionController::Cookiesを追加
- config/application.rbにCookieとSessionのミドルウェアを追加
- database.ymlをdevelopment/testでローカルDB、productionでNeon DBを使うように修正
- test/controllersの不要なテストファイル（brands, passwords）を削除
- test/fixtures/users.ymlにfullnameとadminフィールドを追加
- すべてのテストが正常に通過（4 runs, 9 assertions, 0 failures）
