+++ app/controllers/application_controller.rb
@@ -1,3 +1,4 @@
 class ApplicationController < ActionController::API
+  include ActionController::Cookies
   include Authentication
 end
