+++ config/application.rb
@@ -28,5 +28,9 @@ module FashionTrendsTrackerApi
     # Middleware like session, flash, cookies can be added back manually.
     # Skip views, helpers and assets when generating a new resource.
     config.api_only = true
+    
+    # Cookie と Session を有効化（認証に必要）
+    config.middleware.use ActionDispatch::Cookies
+    config.middleware.use ActionDispatch::Session::CookieStore
   end
 end
