+++ test/fixtures/users.yml
@@ -3,7 +3,11 @@
 one:
   email_address: one@example.com
   password_digest: <%= password_digest %>
+  fullname: "User One"
+  admin: false
 
 two:
   email_address: two@example.com
   password_digest: <%= password_digest %>
+  fullname: "User Two"
+  admin: false
